cook_master/backend/requirements.txt
aniso8601==9.0.1
attrs==22.1.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.0
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
gevent==21.12.0
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==1.4.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
urllib3==1.26.12
uWSGI==2.0.20
Werkzeug==2.0.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0